+++ api/specs/web-server/_users.py
@@ -6,7 +6,6 @@
 
 from typing import Annotated
 
-from _common import Error, Log
 from fastapi import APIRouter, Depends, status
 from models_library.api_schemas_webserver.users_preferences import PatchRequestBody
 from models_library.generics import Envelope
@@ -23,7 +22,6 @@ from simcore_service_webserver.users._notifications import (
 )
 from simcore_service_webserver.users.schemas import (
     PermissionGet,
-    ProfileCredentialsCheck,
     ProfileGet,
     ProfileUpdate,
     Token,
@@ -46,16 +44,6 @@ async def get_my_profile():
     status_code=status.HTTP_204_NO_CONTENT,
 )
 async def update_my_profile(_profile: ProfileUpdate):
-    ...
-
-
-@router.post(
-    "/me:mark-deleted",
-    response_model=Envelope[Log],
-    status_code=status.HTTP_200_OK,
-    responses={status.HTTP_409_CONFLICT: {"model": Envelope[Error]}},
-)
-async def mark_account_for_deletion(_body: ProfileCredentialsCheck):
     ...
