+++ packages/models-library/src/models_library/api_schemas_webserver/auth.py
@@ -1,0 +1,9 @@
+from models_library.emails import LowerCaseEmailStr
+from pydantic import SecretStr
+
+from ._base import InputSchema
+
+
+class UnregisterCheck(InputSchema):
+    email: LowerCaseEmailStr
+    password: SecretStr
