+++ services/web/server/src/simcore_service_webserver/users/api.py
@@ -23,6 +23,7 @@ from ..groups.models import convert_groups_db_to_schema
 from ..login.storage import AsyncpgStorage, get_plugin_storage
 from ..security.api import clean_auth_policy_cache
 from . import _db
+from ._api import get_user_credentials, set_user_as_deleted
 from ._preferences_api import get_frontend_user_preferences_aggregation
 from .exceptions import UserNotFoundError
 from .schemas import ProfileGet, ProfileUpdate, convert_user_db_to_schema
@@ -265,3 +266,12 @@ async def get_users_in_group(app: web.Application, gid: GroupID) -> set[UserID]:
 async def update_expired_users(engine: Engine) -> list[UserID]:
     async with engine.acquire() as conn:
         return await _db.do_update_expired_users(conn)
+
+
+assert set_user_as_deleted  # nosec
+assert get_user_credentials  # nosec
+
+__all__: tuple[str, ...] = (
+    "get_user_credentials",
+    "set_user_as_deleted",
+)
