+++ services/web/server/src/simcore_service_webserver/users/schemas.py
@@ -8,7 +8,7 @@ from models_library.api_schemas_webserver.groups import AllUsersGroups
 from models_library.api_schemas_webserver.users_preferences import AggregatedPreferences
 from models_library.basic_types import IdInt
 from models_library.emails import LowerCaseEmailStr
-from pydantic import BaseModel, Field, SecretStr, validator
+from pydantic import BaseModel, Field, validator
 from servicelib.json_serialization import json_dumps
 from simcore_postgres_database.models.users import UserRole
 
@@ -66,11 +66,6 @@ class _ProfileCommon(BaseModel):
 
 class ProfileUpdate(_ProfileCommon):
     pass
-
-
-class ProfileCredentialsCheck(BaseModel):
-    email: LowerCaseEmailStr
-    password: SecretStr
 
 
 class ProfileGet(_ProfileCommon):
